+++ MangoAPI.DataAccess/Database/Configurations/UserChatEntityConfiguration.cs
@@ -10,7 +10,7 @@ namespace MangoAPI.DataAccess.Database.Configurations
     {
         public void Configure(EntityTypeBuilder<UserChatEntity> builder)
         {
-            builder.HasKey(x => new { x.ChatId, x.UserId });
+            builder.HasKey(x => new {x.ChatId, x.UserId});
             builder.Property(x => x.RoleId).IsRequired();
 
             builder.HasOne(x => x.Chat)
@@ -26,187 +26,187 @@ namespace MangoAPI.DataAccess.Database.Configurations
                 {
                     UserId = SeedDataConstants.PetroId,
                     ChatId = SeedDataConstants.WsbId,
-                    RoleId = UserRole.Moderator,
+                    RoleId = (int) UserRole.Moderator,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.SzymonId,
                     ChatId = SeedDataConstants.WsbId,
-                    RoleId = UserRole.Owner,
+                    RoleId = (int) UserRole.Owner,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.IlliaId,
                     ChatId = SeedDataConstants.WsbId,
-                    RoleId = UserRole.User,
+                    RoleId = (int) UserRole.User,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.ArslanbekId,
                     ChatId = SeedDataConstants.WsbId,
-                    RoleId = UserRole.User,
+                    RoleId = (int) UserRole.User,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.SerhiiId,
                     ChatId = SeedDataConstants.WsbId,
-                    RoleId = UserRole.User,
+                    RoleId = (int) UserRole.User,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.KhachaturId,
                     ChatId = SeedDataConstants.ExtremeCodeMainId,
-                    RoleId = UserRole.User,
+                    RoleId = (int) UserRole.User,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.RazumovskyId,
                     ChatId = SeedDataConstants.ExtremeCodeMainId,
-                    RoleId = UserRole.Admin,
+                    RoleId = (int) UserRole.Admin,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.KolbasatorId,
                     ChatId = SeedDataConstants.ExtremeCodeMainId,
-                    RoleId = UserRole.Moderator,
+                    RoleId = (int) UserRole.Moderator,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.AmelitId,
                     ChatId = SeedDataConstants.ExtremeCodeMainId,
-                    RoleId = UserRole.Owner,
+                    RoleId = (int) UserRole.Owner,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.RazumovskyId,
                     ChatId = SeedDataConstants.ExtremeCodeFloodId,
-                    RoleId = UserRole.Owner,
+                    RoleId = (int) UserRole.Owner,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.KolbasatorId,
                     ChatId = SeedDataConstants.ExtremeCodeFloodId,
-                    RoleId = UserRole.Moderator,
+                    RoleId = (int) UserRole.Moderator,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.KhachaturId,
                     ChatId = SeedDataConstants.ExtremeCodeFloodId,
-                    RoleId = UserRole.User,
+                    RoleId = (int) UserRole.User,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.AmelitId,
                     ChatId = SeedDataConstants.ExtremeCodeFloodId,
-                    RoleId = UserRole.User,
+                    RoleId = (int) UserRole.User,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.AmelitId,
                     ChatId = SeedDataConstants.ExtremeCodeCppId,
-                    RoleId = UserRole.Owner,
+                    RoleId = (int) UserRole.Owner,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.RazumovskyId,
                     ChatId = SeedDataConstants.ExtremeCodeCppId,
-                    RoleId = UserRole.Admin,
+                    RoleId = (int) UserRole.Admin,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.KhachaturId,
                     ChatId = SeedDataConstants.ExtremeCodeCppId,
-                    RoleId = UserRole.User,
+                    RoleId = (int) UserRole.User,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.KolbasatorId,
                     ChatId = SeedDataConstants.ExtremeCodeCppId,
-                    RoleId = UserRole.User,
+                    RoleId = (int) UserRole.User,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.RazumovskyId,
                     ChatId = SeedDataConstants.ExtremeCodeDotnetId,
-                    RoleId = UserRole.Owner,
+                    RoleId = (int) UserRole.Owner,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.KolbasatorId,
                     ChatId = SeedDataConstants.ExtremeCodeDotnetId,
-                    RoleId = UserRole.User,
+                    RoleId = (int) UserRole.User,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.KhachaturId,
                     ChatId = SeedDataConstants.ExtremeCodeDotnetId,
-                    RoleId = UserRole.User,
+                    RoleId = (int) UserRole.User,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.AmelitId,
                     ChatId = SeedDataConstants.ExtremeCodeDotnetId,
-                    RoleId = UserRole.User,
+                    RoleId = (int) UserRole.User,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.KhachaturId,
                     ChatId = SeedDataConstants.DirectKhachaturRazumovsky,
-                    RoleId = UserRole.User,
+                    RoleId = (int) UserRole.User,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.RazumovskyId,
                     ChatId = SeedDataConstants.DirectKhachaturRazumovsky,
-                    RoleId = UserRole.User,
+                    RoleId = (int) UserRole.User,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.KolbasatorId,
                     ChatId = SeedDataConstants.DirectKolbasatorRazumovsky,
-                    RoleId = UserRole.User,
+                    RoleId = (int) UserRole.User,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.RazumovskyId,
                     ChatId = SeedDataConstants.DirectKolbasatorRazumovsky,
-                    RoleId = UserRole.User,
+                    RoleId = (int) UserRole.User,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.AmelitId,
                     ChatId = SeedDataConstants.DirectAmelitRazumovsky,
-                    RoleId = UserRole.User,
+                    RoleId = (int) UserRole.User,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.RazumovskyId,
                     ChatId = SeedDataConstants.DirectAmelitRazumovsky,
-                    RoleId = UserRole.User,
+                    RoleId = (int) UserRole.User,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.KhachaturId,
                     ChatId = SeedDataConstants.DirectKhachaturKolbasator,
-                    RoleId = UserRole.User,
+                    RoleId = (int) UserRole.User,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.KolbasatorId,
                     ChatId = SeedDataConstants.DirectKhachaturKolbasator,
-                    RoleId = UserRole.User,
+                    RoleId = (int) UserRole.User,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.PetroId,
                     ChatId = SeedDataConstants.DirectPetroSzymon,
-                    RoleId = UserRole.User,
+                    RoleId = (int) UserRole.User,
                 },
                 new UserChatEntity
                 {
                     UserId = SeedDataConstants.SzymonId,
                     ChatId = SeedDataConstants.DirectPetroSzymon,
-                    RoleId = UserRole.User,
+                    RoleId = (int) UserRole.User,
                 });
         }
     }
